buildsystem: use 16k max page size
By passing "-z max-page-size=16384" to the linker.

Newer Android devices use 16KB and require a recompilation with this flag.
See https://developer.android.com/guide/practices/page-sizes

Neither VLC nor the contribs (used by Android) use PAGE_SIZE directly so it
should be fine.

This flag is needed regardless the use of the NDK 27 or not.

The CFLAGS/CXXFLAGS is probably not needed by added just in case.