buildsystem: add an option to build C++ code with static runtime
By default we use the dynamic runtime and need to ship libc++_shared.so.

This is also needed in libvlcjni because ndk-build copies the libc++_shared.so
depending on the value of APP_STL. So even if it's not copied for libvlc.so
it would always be copied with libvlcjni.so.

Partially reverts e45d311dbbfd835f0ae3beed117229e779d69dc5.

The libvlcjni Java code still loads the c++_shared library unconditionally.
We could use ReLink and let the loader do its job [^1].
Apparently that's what happens since API 23.

[^1]: https://github.com/KeepSafe/ReLinker